Adjusts Installer building rules to avoid inconsistent reinstalls.
If an user installed Bridge to a directory A then reinstalled it to a
directory B without previously uninstalling A's, then uninstall would
lose information about installation under A.

This makes it so when reinstalling, the installer won't allow the user
to change the destination directory and will only use the one saved in
the system registry.

Information will be displaying telling the user to uninstall bridge
before reinstalling to a different folder.