Sets up installer to fully support either global or local installs.
The default is hardcoded local install. Ideal would be to have different
installer files for local or global installs, as complexity of the
installer definition file will get much higher if support for user
choice (install to current user or whole host) is implemented.

Current user install mode does not require administrative privileges in
order to install bridge. Allowing the user to choose would allow the user
to install both current user or whole machine and more checks and support
for handling the different installations would be required.