Reviews NSIS install process.
Basically adds a backup registry key for the PATH environment variable.
If the backup path is present, during installation, the user will be
asked whether they want to wipe the backup path key as well, or keep it
(in case they are sure their path is not broken, they can be sure
there's nothing left from Bridge's installation).

This is done in an attempt to improve the situation that happens around
issue bridgedotnet/CLI#15.